refactor(mcp): extract browser-agnostic gateway core
