+++ .github/workflows/mcp-quality.yml
@@ -3,6 +3,8 @@ name: MCP Quality Gate
 on:
   pull_request:
     paths:
+      - 'packages/shared/lib/mcp-core/**'
+      - 'packages/shared/index.mts'
       - 'pages/content/src/core/**'
       - 'pages/content/src/components/sidebar/Instructions/**'
       - 'pages/content/src/stores/tool.store.ts'
@@ -71,6 +73,13 @@ jobs:
       - name: Check MCP formatting
         run: >-
           pnpm exec prettier --check
+          packages/shared/lib/mcp-core/context-budget.ts
+          packages/shared/lib/mcp-core/tool-router.ts
+          packages/shared/lib/mcp-core/execution-policy.ts
+          packages/shared/lib/mcp-core/mcp-telemetry.ts
+          packages/shared/lib/mcp-core/gateway.ts
+          packages/shared/lib/mcp-core/index.ts
+          packages/shared/index.mts
           pages/content/src/core/context-budget.ts
           pages/content/src/core/tool-router.ts
           pages/content/src/core/execution-policy.ts
