+++ packages/shared/index.mts
@@ -1,3 +1,4 @@
 export * from './lib/hooks/index.js';
 export * from './lib/hoc/index.js';
 export * from './lib/utils/index.js';
+export * from './lib/mcp-core/index.js';
