+++ packages/shared/lib/mcp-core/index.ts
@@ -1,0 +1,5 @@
+export * from './context-budget.js';
+export * from './execution-policy.js';
+export * from './gateway.js';
+export * from './mcp-telemetry.js';
+export * from './tool-router.js';
